feat: 메인페이지 화면수정 - 배너 크기 변경 및 버튼 추가 - 검색박스 사이즈에 따른 수정

+++ src/main/java/environment/project/dto/ClubApplyDTO.java
@@ -6,8 +6,8 @@ import java.sql.Date;
 
 @Data
 public class ClubApplyDTO {
-    private Long clubApplyNo;
-    private Long userNo;
+    private int clubApplyNo;
+    private int userNo;
     private String clubName;
     private String clubPurpose;
     private String clubActivities;
